Implement Chinese(Taiwan) variant
This patch implements the Chinese(Taiwan) variant of the snap store
badge.

The localized prompt text is literally "(If you want )to install the
application please refer to the", the direct translation isn't possible
as it requires text both before and after the store brand:

    自
    Snap Store
    取得軟體

so some alternations has to be done.

The Chinese font used for the Chinese text is "Noto Sans CJK TC" which
can be acquired by installing the fonts-noto-cjk package(likely shipped
by default).

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>